refactor: extract starter bootstrap command family
