+++ loopx/cli_commands/__init__.py
@@ -51,8 +51,6 @@ from .registry_admin import (
 )
 from .starter import (
     handle_codex_cli_bounded_visible_pilot_adapter_command,
-    handle_codex_cli_bootstrap_message_command,
-    handle_codex_cli_exec_handoff_command,
     handle_codex_cli_visible_first_response_capture_plan_command,
     handle_codex_cli_local_driver_plan_command,
     handle_codex_cli_local_scheduler_exec_command,
@@ -60,16 +58,22 @@ from .starter import (
     handle_codex_cli_one_message_loop_pilot_command,
     handle_codex_cli_runtime_idle_detector_command,
     handle_codex_cli_session_probe_command,
-    handle_codex_cli_tui_bootstrap_smoke_bundle_command,
     handle_codex_cli_visible_attach_acceptance_command,
     handle_codex_cli_visible_local_driver_pilot_command,
     handle_codex_cli_visible_driver_run_command,
     handle_codex_cli_visible_driver_plan_command,
     handle_codex_cli_visible_session_proof_command,
     handle_demo_command,
-    handle_new_project_prompt_command,
     handle_starter_command,
     register_starter_commands,
+)
+from .starter_bootstrap import (
+    handle_codex_cli_bootstrap_message_command,
+    handle_codex_cli_exec_handoff_command,
+    handle_codex_cli_tui_bootstrap_smoke_bundle_command,
+    handle_new_project_prompt_command,
+    handle_starter_bootstrap_command,
+    register_starter_bootstrap_commands,
 )
 from .status import (
     handle_check_command,
@@ -131,6 +135,7 @@ __all__ = [
     "handle_review_packet_command",
     "handle_status_command",
     "handle_starter_command",
+    "handle_starter_bootstrap_command",
     "handle_support_control_command",
     "handle_terminal_bench_adapter_command",
     "handle_terminal_bench_environment_result_command",
@@ -151,6 +156,7 @@ __all__ = [
     "register_quota_command",
     "register_registry_admin_commands",
     "register_starter_commands",
+    "register_starter_bootstrap_commands",
     "register_status_commands",
     "register_support_control_commands",
     "register_terminal_bench_adapter_commands",
