+++ loopx/cli_commands/starter.py
@@ -13,18 +13,6 @@ from ..demo import (
     render_demo_markdown,
     run_demo,
 )
-from ..project_prompt import (
-    DEFAULT_HANDOFF_ADAPTER_KIND,
-    DEFAULT_HANDOFF_ADAPTER_STATUS,
-    build_codex_cli_bootstrap_message,
-    build_codex_cli_exec_handoff,
-    build_codex_cli_tui_bootstrap_smoke_bundle,
-    build_new_project_prompt,
-    render_codex_cli_bootstrap_message_markdown,
-    render_codex_cli_exec_handoff_markdown,
-    render_codex_cli_tui_bootstrap_smoke_bundle_markdown,
-    render_new_project_prompt_markdown,
-)
 from ..codex_cli_probe import (
     DEFAULT_CODEX_BIN,
     DEFAULT_EXECUTOR_TIMEOUT_SECONDS,
@@ -61,6 +49,10 @@ from ..codex_cli_probe import (
     render_codex_cli_runtime_idle_detector_markdown,
     probe_human_input_idle_seconds,
     run_codex_cli_session_probe,
+)
+from .starter_bootstrap import (
+    handle_starter_bootstrap_command,
+    register_starter_bootstrap_commands,
 )
 
 
@@ -167,67 +159,7 @@ def _load_codex_cli_runtime_idle_payload(args: argparse.Namespace) -> dict[str, 
 
 
 def register_starter_commands(subparsers: argparse._SubParsersAction) -> None:
-    prompt_parser = subparsers.add_parser(
-        "new-project-prompt",
-        help="Generate a copy-paste Codex prompt for connecting a project from a goal document.",
-    )
-    prompt_parser.add_argument("--project", required=True, help="Project directory the target Codex session can access.")
-    prompt_parser.add_argument("--goal-doc", required=True, help="Goal document path for the target project.")
-    prompt_parser.add_argument("--goal-id", help="Initial stable goal id. Defaults to <project-name>-goal.")
-    prompt_parser.add_argument("--objective", help="Initial objective. Defaults to an extraction placeholder.")
-    prompt_parser.add_argument("--domain", help="Initial domain label. Defaults to an extraction placeholder.")
-    prompt_parser.add_argument("--adapter-kind", default=DEFAULT_HANDOFF_ADAPTER_KIND)
-    prompt_parser.add_argument("--adapter-status", default=DEFAULT_HANDOFF_ADAPTER_STATUS)
-    prompt_parser.add_argument("--next-probe", help="Optional read-only pre-tick command for the target project.")
-    prompt_parser.add_argument("--spawn-allowed", action="store_true", help="Include controller/sub-agent flags.")
-    prompt_parser.add_argument("--allowed-domain", action="append", default=[], help="Allowed child work domain. Repeatable.")
-    prompt_parser.add_argument("--write-scope", action="append", default=[], help="Allowed write scope such as docs/**. Repeatable.")
-
-    codex_cli_bootstrap_parser = subparsers.add_parser(
-        "codex-cli-bootstrap-message",
-        help="Generate a one-message Codex CLI TUI setup prompt that installs/connects LoopX, then sets the thin /goal loop.",
-    )
-    codex_cli_bootstrap_parser.add_argument("--project", default=".", help="Project directory to start from.")
-    codex_cli_bootstrap_parser.add_argument("--goal-id", help="Goal id. Defaults to <project-name>-goal.")
-    codex_cli_bootstrap_parser.add_argument(
-        "--agent-id",
-        help="Registered LoopX agent id to include in quota/claim instructions.",
-    )
-    codex_cli_bootstrap_parser.add_argument(
-        "--cli-bin",
-        default="loopx",
-        help="LoopX CLI binary name embedded in generated commands.",
-    )
-    codex_cli_bootstrap_parser.add_argument(
-        "--message-only",
-        "--copy-only",
-        dest="message_only",
-        action="store_true",
-        help="Print only the Codex CLI TUI paste message, without the Markdown review packet.",
-    )
-
-    codex_cli_tui_bootstrap_smoke_parser = subparsers.add_parser(
-        "codex-cli-tui-bootstrap-smoke-bundle",
-        help="Generate a transcript-free first-run smoke packet for the Codex CLI TUI bootstrap path.",
-    )
-    codex_cli_tui_bootstrap_smoke_parser.add_argument(
-        "--project",
-        default=".",
-        help="Project directory to model as the fresh repo.",
-    )
-    codex_cli_tui_bootstrap_smoke_parser.add_argument(
-        "--goal-id",
-        help="Goal id. Defaults to <project-name>-goal.",
-    )
-    codex_cli_tui_bootstrap_smoke_parser.add_argument(
-        "--agent-id",
-        help="Registered LoopX agent id to include in quota/claim instructions.",
-    )
-    codex_cli_tui_bootstrap_smoke_parser.add_argument(
-        "--cli-bin",
-        default="loopx",
-        help="LoopX CLI binary name embedded in generated commands.",
-    )
+    register_starter_bootstrap_commands(subparsers)
 
     codex_cli_one_message_loop_parser = subparsers.add_parser(
         "codex-cli-one-message-loop-pilot",
@@ -737,27 +669,6 @@ def register_starter_commands(subparsers: argparse._SubParsersAction) -> None:
         help="Confirm manual takeover remains available.",
     )
 
-    codex_cli_exec_parser = subparsers.add_parser(
-        "codex-cli-exec-handoff",
-        help="Show the disabled headless handoff boundary and the TUI message-only bootstrap command.",
-    )
-    codex_cli_exec_parser.add_argument("--project", default=".", help="Project directory to start from.")
-    codex_cli_exec_parser.add_argument("--goal-id", help="Goal id. Defaults to <project-name>-goal.")
-    codex_cli_exec_parser.add_argument(
-        "--agent-id",
-        help="Registered LoopX agent id to include in quota/claim instructions.",
-    )
-    codex_cli_exec_parser.add_argument(
-        "--cli-bin",
-        default="loopx",
-        help="LoopX CLI binary name embedded in generated commands.",
-    )
-    codex_cli_exec_parser.add_argument(
-        "--codex-bin",
-        default=DEFAULT_CODEX_BIN,
-        help="Codex CLI executable name embedded in the generated handoff command.",
-    )
-
     demo_parser = subparsers.add_parser(
         "demo",
         help="Create a disposable local demo goal and show status/quota output.",
@@ -771,58 +682,6 @@ def register_starter_commands(subparsers: argparse._SubParsersAction) -> None:
     demo_parser.add_argument("--objective", default=DEFAULT_DEMO_OBJECTIVE)
     demo_parser.add_argument("--user-todo", default=DEFAULT_DEMO_USER_TODO)
     demo_parser.add_argument("--agent-todo", default=DEFAULT_DEMO_AGENT_TODO)
-
-
-def handle_new_project_prompt_command(
-    args: argparse.Namespace,
-    print_payload: PrintPayload,
-) -> int:
-    payload = build_new_project_prompt(
-        project=Path(args.project),
-        goal_doc=Path(args.goal_doc),
-        goal_id=args.goal_id,
-        objective=args.objective,
-        domain=args.domain,
-        adapter_kind=args.adapter_kind,
-        adapter_status=args.adapter_status,
-        next_probe=args.next_probe,
-        spawn_allowed=bool(args.spawn_allowed),
-        allowed_domains=args.allowed_domain,
-        write_scope=args.write_scope,
-    )
-    print_payload(payload, args.format, render_new_project_prompt_markdown)
-    return 0
-
-
-def handle_codex_cli_bootstrap_message_command(
-    args: argparse.Namespace,
-    print_payload: PrintPayload,
-) -> int:
-    payload = build_codex_cli_bootstrap_message(
-        project=Path(args.project),
-        goal_id=args.goal_id,
-        agent_id=args.agent_id,
-        cli_bin=args.cli_bin,
-    )
-    if bool(getattr(args, "message_only", False)):
-        print(str(payload.get("message") or ""))
-        return 0
-    print_payload(payload, args.format, render_codex_cli_bootstrap_message_markdown)
-    return 0
-
-
-def handle_codex_cli_tui_bootstrap_smoke_bundle_command(
-    args: argparse.Namespace,
-    print_payload: PrintPayload,
-) -> int:
-    payload = build_codex_cli_tui_bootstrap_smoke_bundle(
-        project=Path(args.project),
-        goal_id=args.goal_id,
-        agent_id=args.agent_id,
-        cli_bin=args.cli_bin,
-    )
-    print_payload(payload, args.format, render_codex_cli_tui_bootstrap_smoke_bundle_markdown)
-    return 0
 
 
 def handle_codex_cli_one_message_loop_pilot_command(
@@ -1136,21 +995,6 @@ def handle_codex_cli_runtime_idle_detector_command(
     return 0 if payload.get("ok") else 1
 
 
-def handle_codex_cli_exec_handoff_command(
-    args: argparse.Namespace,
-    print_payload: PrintPayload,
-) -> int:
-    payload = build_codex_cli_exec_handoff(
-        project=Path(args.project),
-        goal_id=args.goal_id,
-        agent_id=args.agent_id,
-        cli_bin=args.cli_bin,
-        codex_bin=args.codex_bin,
-    )
-    print_payload(payload, args.format, render_codex_cli_exec_handoff_markdown)
-    return 0 if payload.get("ok") else 1
-
-
 def handle_demo_command(args: argparse.Namespace, print_payload: PrintPayload) -> int:
     try:
         payload = run_demo(
@@ -1176,16 +1020,15 @@ def handle_starter_command(
     args: argparse.Namespace,
     print_payload: PrintPayload,
 ) -> int | None:
+    bootstrap_result = handle_starter_bootstrap_command(args, print_payload)
+    if bootstrap_result is not None:
+        return bootstrap_result
     handlers: dict[str, Callable[[argparse.Namespace, PrintPayload], int]] = {
-        "new-project-prompt": handle_new_project_prompt_command,
-        "codex-cli-bootstrap-message": handle_codex_cli_bootstrap_message_command,
-        "codex-cli-tui-bootstrap-smoke-bundle": handle_codex_cli_tui_bootstrap_smoke_bundle_command,
         "codex-cli-one-message-loop-pilot": handle_codex_cli_one_message_loop_pilot_command,
         "codex-cli-visible-local-driver-pilot": handle_codex_cli_visible_local_driver_pilot_command,
         "codex-cli-bounded-visible-pilot-adapter": handle_codex_cli_bounded_visible_pilot_adapter_command,
         "codex-cli-visible-first-response-capture-plan": handle_codex_cli_visible_first_response_capture_plan_command,
         "codex-cli-visible-attach-acceptance": handle_codex_cli_visible_attach_acceptance_command,
-        "codex-cli-exec-handoff": handle_codex_cli_exec_handoff_command,
         "codex-cli-session-probe": handle_codex_cli_session_probe_command,
         "codex-cli-visible-driver-plan": handle_codex_cli_visible_driver_plan_command,
         "codex-cli-local-driver-plan": handle_codex_cli_local_driver_plan_command,
